+++ resources/integration_tests/aws_autoscaling_launch_configurations_test.go
@@ -9,7 +9,7 @@ import (
 	providertest "github.com/cloudquery/cq-provider-sdk/provider/testing"
 )
 
-func TestIntegrationDirectAutoscalingLaunchConfigurations(t *testing.T) {
+func TestIntegrationAutoscalingLaunchConfigurations(t *testing.T) {
 	awsTestIntegrationHelper(t, resources.AutoscalingLaunchConfigurations(), nil, func(res *providertest.ResourceIntegrationTestData) providertest.ResourceIntegrationVerification {
 		return providertest.ResourceIntegrationVerification{
 			Name: "aws_autoscaling_launch_configurations",